Compute front and rear tire pressures based on data for bike and rider.

The label for each point is generated and stored in the `annotation` column.

bike_tire_pressures(rider_weight_lbs = 100, bike_weight_lbs = 15,
  load_lbs = 2, front_tire_casing_compensation = 1,
  front_tire_size_mm = 28, rear_tire_casing_compensation = 1,
  rear_tire_size_mm = 28, front_distribution = 0.4)

Arguments

rider_weight_lbs
Weight of rider dressed with biking shoes and helmet.

bike_weight_lbs
Weight of bike

load_lbs
Weight of accessories carried on the bike such as water bottles, repair kit, pump, etc.

front_tire_casing_compensation
Supple casings require about 15% more pressure to maintain shape.

front_tire_size_mm
Measured or side wall marking.

rear_tire_casing_compensation
Supple casings require about 15% more pressure to maintain shape.

rear_tire_size_mm
Measured or side wall marking.

front_distribution
Decimal representation of percentage of bike plus rider weight on the front wheel. Defaults to 0.4 for dropped handlebar road bike corresponding to 40/60 front/rear weight distribution.

Value

tibble used by ggplot to display tire pressures

Details

For pressures between 105 and 120psi, a suggestion to increase tire width is stored in the `message` column. For pressures over 120psi, a suggestion to check max tire and rim pressures is stored.