btpress

If you were told that you could bicycle faster and more comfortablely simply by changing your bike tire pressure, would you try it? There are now 2 independent studies that support lower pressures are significantly better than pumping up your tires until they are rock hard. See the References section for articles.

Pump Inaccuracies

Pump gauges considerably inaccurate. Parallex viewing makes precise reading of needle impossible. Never mind the pressure sensing mechanism has a wide tolerance.

Rim Widths

The original research was conducted when internal widths for road rims were around 19mm. Today's rims are available in wider sizes, 23 - 25mm, increasing the volume up to 33%. This program does not use internal rim width in its calculations.

Weight vs. Load

At the end of the day, these two are effectively the same. That said, Weight is used for the individual components that go into the calculations. Load is used to describe the forces the tires are subject to.

Directions for Use

Extra light casings need an extra 10-15% pressure to prevent the casing threads from breaking.

There are other subtlies which are not explicitly taken into account. The biggest one is weight carried on the rider. For example, if your bike lock is carried on you, it should be added to the rider weight, not the load field. Same for a backpack. Carrying weight on front and rear racks is not considered by the program.

Snow tires and thick sidewalled tires are supported by their sidewalls and do not respond to pressure changes the way performace tires do.

References

Links to articles:

Science and Bicycles 1: Tires and Pressure

The Tire Pressure Revolution

Tire Pressure: Data and Details

Silca - Road to Roubaix - The Complete Story

Tire Size, Pressure, Aero, Comfort, Rolling Resistance and More. Part 1: How we Got to Now. There five parts in this series.

Bicycle Quarterly article

Road Tire Pressure For Plus And BLACK Wheels

Tire Pressure – Stop Guessing And Read The Science

Acknowledgements

Source for data and pratical application to Jan Heine and Frank Berto published in Bicycle Quarterly.

The emperical curve fitting formula from the BikeTinker.

Full Disclosure

I ride and recommend Compass Bicycle tires and subscribe to Bicycle Quarterly. I received no compensation for this program nor was I asked to write it by anyone.

Disclaimer

Biking riding is an inheritly risky activity. Remember the bike racer's motto: Falling is a matter of 'when', not 'if'. You are solely responsible for not exceeding any limits including, but not exclusive of maximum tire inflation pressure, rim inflation pressure, mininum tire clearance, and compatibility of tubeless tire and rim combinations. Use of this app is at your risk and you assume all liabilities.