btpress

A program to calculate optimized bicycle tire pressure based on research that found a small sag or droop in the profile of the tire lead to better performance and comfort.

Details

Image that you could ride one gear faster or feel less fatigued at the end of your ride and it cost you nothing. Tuning your bike tire pressue to you, your bike, and your load may do just that. It all depends on wider bike tires with flexible sidewalls.

Riding one gear higher, depending on your gearing, is a 5 - 8% increase. There are now 2 independent studies that support lower pressures are significantly better than pumping up your tires until they are rock hard. See the References section for articles.

Directions for Use

The first step is evaluating whether your tires have flexible sidewalls which will respond readily to tire pressure changes. Tires marketed as racing tires, marketed as having a casing from a tubular tire, used by professional racers, or are a folding tire are good candidates. On the other hand, puncture resistant tires, snow tires, and thick sidewalled tires are supported by their sidewalls and do not respond to pressure changes very well.

The simplest way to calculate your tire pressure is to use the companion web app, which uses this package.

To calculate your tire pressures execute this code sequence:
my_bike <- bike_tire_pressures() # Supply your values for the parameters
plot_bike_inflation(bike = my_bike)

If a test ride indicates that the tire pressure is really too low, it's likely your pressure gauge is reading too high. Try increasing your pressure by 10psi. Remember this is not an exact science nor setting so some experimentation may be needed.

Extra light casings need an extra 10-15% pressure to prevent the casing threads from breaking.

Let's say your trials indicate a performance improvement and you'd like to explore other tire options, or it's simply time for new tires. A wider tire is a great option to consider. The following code compares 2 setups and builds upon the first example.
new_tires <- bike_tire_pressures() # Supply new tires for the parameters
plot_2_bike_inflation(bike_a = my_bike, bike_b = new_tires)

There are 2 cases for the load which need to be handled manually. The first is weight carried on the rider. For example, if your bike lock is carried on you, it should be added to the rider weight parameter, not the load parameter. Same for a backpack.

The second is a load on both the front and rear racks. It's unwise to assume a front/rear weight distribution. My recommendation is to measure the distribution on a level surface. A bathroom scale place under one wheel and a spacer block the same height as the scale placed under the other wheel. Load up the bike, hop on, and have a helper steady you and read the numbers. Flip the bike around and meaure the other wheel. Divide the front weight by the rear weight and use that for the front_distribution parameter.

Pump Inaccuracies

Pump gauges considerably inaccurate. Parallex viewing makes precise reading of needle impossible. Never mind the pressure sensing mechanism has a wide tolerance.

Rim Widths

The original research was conducted when internal widths for road rims were around 19mm. Today's rims are available in wider sizes, 23 - 25mm, increasing the volume up to 33%. This program does not use internal rim width in its calculations.

Weight vs. Load

At the end of the day, these two are effectively the same. That said, Weight is used for the individual components that go into the calculations. Load is used to describe the forces the tires are subject to.

References

Science and Bicycles 1: Tires and Pressure

The Tire Pressure Revolution

Tire Pressure: Data and Details

Why Wider Tires Corner Better

Compass Bon Jon Pass 700C x 35 in Top 5 Fastest Tires in the World

Silca - Road to Roubaix - The Complete Story

Part one, Tire Size, Pressure, Aero, Comfort, Rolling Resistance and More. Part 1: How We Got to Now, of a five part series.

The original Bicycle Quarterly article

Road Tire Pressure For Plus And BLACK Wheels

Tire Pressure – Stop Guessing And Read The Science

Acknowledgements

Source for data and pratical application to Jan Heine and Frank Berto published in Bicycle Quarterly.

The emperical curve fitting formula from the BikeTinker.

Disclaimer

Biking riding is an inheritly risky activity. Remember the bike racer's motto: Falling is a matter of 'when', not 'if'. You are solely responsible for not exceeding any limits including, but not exclusive of your abilities, maximum tire inflation pressure, rim inflation pressure, mininum tire clearance, and compatibility of tubeless tire and rim combinations. Use of this app is at your risk and you assume all liabilities.

Full Disclosure

I ride and recommend Compass Bicycle tires and subscribe to Bicycle Quarterly. I received no compensation for this program.